add WIP day 9 solution
Feeling confident about the solution design, but running into memory
issues or something ... can't copy the preamble. Output (preamble
length: 5):

address of preamble: 0x7ffeebd996b0,
address of nums: 0x7ffeebd996b0,
address of buf: 0x7ffeebd97750
first n nums:
17
42
18
39
1
preamble to check:
17
0
0
0

First, check the addresses of the start of `preamble` and the end
of `nums`. Maybe `memcpy` detects that the bounds overlap and stops?

The `nums` array is initialized properly, but not all the numbers
are copied into `preamble`. Could also be misuse of `memcpy`, but
that should be type-agnostic ...