d9p1: `memset` and `memcpy` use proper size arg
Before this commit, I failed to consider that these functions consume
a number of `bytes`, and that we are moving around `long` primitives.
So we need to update the number of bytes properly.

This commit also fixes an indexing bug -- when we validate a given
`preamble` and want to move on to the next 25 numbers, we only update
the `offset` for the `src` array, `nums`, that we are copying from -
we copy to the same locations in the `preamble` array.